changed int to size_t

+++ sorting/selection_sort.h
@@ -16,9 +16,9 @@ void SelectionSort(std::vector<T> &A, bool order) {
   switch (order) {
     case true:
       
-      for (int i = 0; i < length - 1; i++) {
+      for (size_t i = 0; i < length - 1; i++) {
         int min_idx = i;
-        for (int j = i; j < length; j++) {
+        for (size_t j = i; j < length; j++) {
           if (A[j] < A[min_idx]) {
             min_idx = j;
           }
@@ -29,9 +29,9 @@ void SelectionSort(std::vector<T> &A, bool order) {
       break;
     case false:
 
-      for (int i = 0; i < length - 1; i++) {
+      for (size_t i = 0; i < length - 1; i++) {
         int max_idx = i;
-        for (int j = i; j < length; j++) {
+        for (size_t j = i; j < length; j++) {
           if (A[j] > A[max_idx]) {
             max_idx = j;
           }
